demo03/requirements.txt
crewai
langchain
langchain-community
openai
python-dotenv
pandas
numpy
tiktoken  # 可选但推荐
matplotlib
seaborn